requirements.txt
GitPython==3.1.18
python-gitlab==2.10.1
requests==2.26.0
